// src/app/models/EmployeeProfile.ts
export class EmployeeProfile {
    id: number;
    detailId: number;
    nameFl: string;
    nameSl: string;
    startDate: Date;
    endDate: Date;
    employeeId: number;
    employeeProfile: number;
    currentAdminLevelId: number;
    currentAdminLeveNameFl: string;
    currentAdminLeveNameSl: string;
}